reorganise repo and update docs
