+++ scrapers/AGENTS.md
@@ -1,7 +1,7 @@
 # Folder Overview
 
 All data collection scripts.
-- `universe.py` pulls ticker lists including S&P 400, S&P 500, S&P 600 and Russell 2000.
+- `universe.py` pulls ticker lists for the S&P 400, S&P 500 and Russell 2000.
 - `wiki.py` and others in this folder fetch alternative data from QuiverQuant and public APIs.
 
 Scrapers call `init_db()` to ensure tables exist and the `universe` helper stores index constituents to Postgres and CSV.
