+++ scrapers/wiki.py
@@ -2,7 +2,7 @@ import datetime as dt
 import json
 from typing import List
 
-from config import QUIVER_RATE_SEC
+from service.config import QUIVER_RATE_SEC
 from infra.rate_limiter import DynamicRateLimiter
 from infra.smart_scraper import get as scrape_get
 from database import db, pf_coll, wiki_coll, init_db
