+++ ws/AGENTS.md
@@ -3,6 +3,6 @@
 WebSocket hub used for push updates to the frontend.
 - `hub.py` handles active connections and broadcast messages.
 
-Depends on `api.py` to orchestrate portfolio events.
+Depends on the API module to orchestrate portfolio events.
 Recent commits added automatic connection metrics so clients can monitor feed
 health via Prometheus.
